Same K and r as for spatstat.
Code was modularized for clarity.

We now support arbitrary, given radius vectors. By
default the r vector is the same as the one from
spatstat.

We give the same result as spatstat if the
1 / \lambda^2 estimate is changed for the
unbiased one. The last bin still differs so that
our K estimate has more mass in the last bin than
the estimate from spatstat.